Move CornerRadii into its own file
